requirements.txt
flask==2.3.3
pillow==10.0.0
transformers==4.42.2
torch==2.1.0
openai==1.0.0
gunicorn
